+++ sjkscan/postprocessing.py
@@ -1,11 +1,26 @@
+# -*- coding: utf-8 -*-
+"""
+    sjkscan.postprocessing
+    ~~~~~~~~~~~~~~~~~~~~~~
+
+    Implements all post processing related actions that sjkscan take on a
+    scanned document.
+
+    :copyright: (c) 2016 by Svante Kvarnström
+    :license: BSD, see LICENSE for more details.
+"""
+
+import logging
 import os
 import re
 import time
 
-from .config import config, load_config
-from .utils import run_cmd, files, move, remove
 from PyPDF2 import PdfFileMerger
 from wand.image import Image
+
+from .config import config, load_config
+from .logger import init_logging
+from .utils import run_cmd, files, move, remove, is_scan_name, parse_args
 
 
 def rotate_image(filename, degrees):
@@ -15,6 +30,7 @@ def rotate_image(filename, degrees):
     :param int degrees: amount of degrees to rotate
 
     """
+    logging.info('Rotating %s %s degrees', filename, degrees)
     with Image(filename=filename) as image:
         with image.clone() as rotated:
             rotated.rotate(degrees)
@@ -28,6 +44,7 @@ def rotate_all_images_in_dir(dirname, degrees):
     :param int degrees: number of degrees to rotate
 
     """
+    logging.info('Rotating images %s degrees in directory %s', dirname, degrees)
     for f in files(dirname):
         rotate_image(os.path.join(dirname, f), degrees)
 
@@ -35,9 +52,10 @@ def rotate_all_images_in_dir(dirname, degrees):
 def unpaper(filename):
     """Process file with unpaper and delete original.
 
-    :param filename: TODO
+    :param filename: file to run unpaper on
 
     """
+    logging.info('Running unpaper on %s', filename)
     unpapered_filename = filename + '.unpapered'
     # TODO: We don't use unpaper's --overwrite because it currently seems to be
     # broken. Once it's been fixed, just --overwrite the original.
@@ -52,7 +70,6 @@ def unpaper_dir(directory, extension=None):
     :param string extension: extension of files to run unpaper on
 
     """
-
     for f in files(directory, extension):
         unpaper(os.path.join(directory, f))
 
@@ -71,6 +88,7 @@ def is_blank(filename):
 
     """
     if not os.path.exists(filename):
+        logging.debug('is_blank: file %s does not exist.')
         return True
 
     c = 'identify -verbose %s' % filename
@@ -83,8 +101,10 @@ def is_blank(filename):
         if match:
             stdev = float(match.group('percent'))
             if stdev > 0.05:
+                logging.debug('is_blank: %s is NOT blank - standard deviation > 0.05 (%d)', filename, stdev)
                 return False
 
+    logging.debug('is_blank: %s is probably blank', filename)
     return True
 
 
@@ -124,7 +144,6 @@ def remove_if_blank(filename):
 
     """
     if is_blank(filename):
-        print('Removing (probably) blank page {}'.format(filename))
         remove(filename)
 
 
@@ -140,16 +159,18 @@ def merge_pdfs(inputs, output):
 
     out = open(output, 'wb')
 
+    logging.info('Merging PDF files into %s...', output)
+
     for filename in inputs:
-        # TODO: This should be a logger call.
-        print('Merging {} -> {}'.format(filename, output))
+        logging.debug('Merging %s -> %s', filename, output)
         try:
             input_fds[filename] = open(filename, 'rb')
         except OSError as e:
-            print('Error opening {}: {}'.format(filename, e))
+            logging.error('Could not open %s: %s', filename, e)
         merger.append(input_fds[filename])
 
     merger.write(out)
+    logging.info('Finished merging PDF files into %s', output)
 
 
 def merge_pdfs_in_dir(directory, output):
@@ -173,6 +194,7 @@ def ocr(filename, language):
     :param string language: language(s) expected to be used in file
 
     """
+    logging.info('Performing OCR (%s) on %s', language, filename)
     base_output_name = filename[:-4]
     command = 'tesseract {} {} -l {} pdf'.format(filename,
                                                  base_output_name,
@@ -191,7 +213,7 @@ def ocr_pnms_in_dir(directory, language):
         ocr(os.path.join(directory, file), language)
 
 
-def scand():
+def main(argv=None):
     """
     Polls DATA_DIR for finished scans. Once found, scand will:
 
@@ -204,18 +226,17 @@ def scand():
     """
 
     load_config()
+    args = parse_args(argv)
+    init_logging(config['Logging']['level'])
+
 
     while True:
         for entry in os.scandir(config['Paths']['data']):
+            if not entry.is_dir() or not is_scan_name(entry.name):
+                continue
+
             archive_dir = config['Paths']['archive']
             inbox_dir = config['Paths']['inbox']
-
-            if entry.name.endswith('.unfinished') or not entry.is_dir():
-                continue
-
-            if entry.name in [os.path.basename(archive_dir), os.path.basename(inbox_dir)]:
-                continue
-
             scan_dir = os.path.join(config['Paths']['data'], entry.name)
             pdf_output = os.path.join(inbox_dir, '{}.pdf'.format(entry.name))
             blank_dir = os.path.join(scan_dir, 'blank')