+++ tests/tests_utils.py
@@ -6,7 +6,9 @@ import shutil
 import tempfile
 import unittest
 
-from sjkscan.utils import files, move, remove
+from sjkscan.utils import files, move, remove, is_scan_name, version
+from sjkscan.config import load_config
+from sjkscan import __version__
 
 
 class TestFiles(unittest.TestCase):
@@ -82,6 +84,22 @@ class TestRemove(unittest.TestCase):
 
     def tearDown(self):
         shutil.rmtree(self.temp_dir)
+
+
+class TestIsScanName(unittest.TestCase):
+    def setUp(self):
+        load_config()
+
+    def test_correct_scan_name_should_be_true(self):
+        self.assertTrue(is_scan_name('2015-02-23_14-13-12'))
+
+    def test_incorrect_scan_name_should_be_false(self):
+        self.assertFalse(is_scan_name('blah'))
+
+
+class TestVersion(unittest.TestCase):
+    def test_version_returns_correct_version(self):
+        self.assertEqual(version(), __version__)
 
 
 if __name__ == "__main__":