+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+- Utility function is_scan_name(name) which returns True if name matches config dir_format and otherwise False
+- -V and --version now prints version and exits.
+- Lots of debug/info logging
+- Log level is settable through sjkscan.conf and command line switches -l and --loglevel
+
+### Changed
+- scand will now only process directories with names corresponding to dir_format
+
 ## [1.1.0] - 2016-02-25
 
 ### Added
@@ -15,3 +26,6 @@
 
 ### Fixed
 - Configuration file interpolation. Setting inbox to '%(data)s/INBOX' in the config file will now expand to /whatever/data/is/INBOX.
+
+[Unreleased]: https://github.com/sjktje/sjkscan/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/sjktje/sjkscan/compare/v1.0.0...v1.1.0